+++ StellarDotnetSdk.Tests/Sep/Sep0009/KycJsonSerializationTest.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Text.Json;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -18,13 +17,13 @@ public class KycJsonSerializationTest
     public void NaturalPersonKycFields_JsonRoundTrip_PreservesDateFields()
     {
         var json = """
-            {
-              "firstName": "John",
-              "birthDate": "1990-01-15",
-              "idIssueDate": "2020-05-10",
-              "idExpirationDate": "2030-05-10"
-            }
-            """;
+                   {
+                     "firstName": "John",
+                     "birthDate": "1990-01-15",
+                     "idIssueDate": "2020-05-10",
+                     "idExpirationDate": "2030-05-10"
+                   }
+                   """;
 
         var fields = JsonSerializer.Deserialize<NaturalPersonKycFields>(json, KycJsonOptions.Default);
 
@@ -34,23 +33,21 @@ public class KycJsonSerializationTest
         fields.IdIssueDate.Should().Be(KycTestDates.IdIssueDate);
         fields.IdExpirationDate.Should().Be(KycTestDates.IdExpirationDate);
 
-#if !TEST_SDK_NETSTANDARD21
         var roundTrip = JsonSerializer.Serialize(fields, KycJsonOptions.Default);
         roundTrip.Should().Contain("\"birthDate\":\"1990-01-15\"");
         roundTrip.Should().Contain("\"idIssueDate\":\"2020-05-10\"");
         roundTrip.Should().Contain("\"idExpirationDate\":\"2030-05-10\"");
-#endif
     }
 
     [TestMethod]
     public void OrganizationKycFields_JsonRoundTrip_PreservesRegistrationDate()
     {
         var json = """
-            {
-              "name": "Acme Corp",
-              "registrationDate": "2020-01-15"
-            }
-            """;
+                   {
+                     "name": "Acme Corp",
+                     "registrationDate": "2020-01-15"
+                   }
+                   """;
 
         var fields = JsonSerializer.Deserialize<OrganizationKycFields>(json, KycJsonOptions.Default);
 
@@ -58,25 +55,23 @@ public class KycJsonSerializationTest
         fields!.Name.Should().Be("Acme Corp");
         fields.RegistrationDate.Should().Be(KycTestDates.RegistrationDateIso);
 
-#if !TEST_SDK_NETSTANDARD21
         var roundTrip = JsonSerializer.Serialize(fields, KycJsonOptions.Default);
         roundTrip.Should().Contain("\"registrationDate\":\"2020-01-15\"");
-#endif
     }
 
     [TestMethod]
     public void StandardKycFields_JsonRoundTrip_PreservesNestedDates()
     {
         var json = """
-            {
-              "naturalPerson": {
-                "birthDate": "1985-12-25"
-              },
-              "organization": {
-                "registrationDate": "2010-06-01"
-              }
-            }
-            """;
+                   {
+                     "naturalPerson": {
+                       "birthDate": "1985-12-25"
+                     },
+                     "organization": {
+                       "registrationDate": "2010-06-01"
+                     }
+                   }
+                   """;
 
         var fields = JsonSerializer.Deserialize<StandardKycFields>(json, KycJsonOptions.Default);
 
@@ -85,7 +80,6 @@ public class KycJsonSerializationTest
         fields.Organization!.RegistrationDate.Should().Be(KycTestDates.NestedRegistrationDate);
     }
 
-#if NET8_0_OR_GREATER && !TEST_SDK_NETSTANDARD21
     [TestMethod]
     public void JsonOptions_DefaultOptions_IsReadOnly()
     {
@@ -98,6 +92,7 @@ public class KycJsonSerializationTest
         KycJsonOptions.Default.IsReadOnly.Should().BeTrue();
     }
 
+#if !TEST_SDK_NETSTANDARD21
     [TestMethod]
     public void NullableDateOnlyJsonConverter_RejectsInvalidDateFormat()
     {
@@ -106,6 +101,49 @@ public class KycJsonSerializationTest
         var act = () => JsonSerializer.Deserialize<NaturalPersonKycFields>(json, KycJsonOptions.Default);
 
         act.Should().Throw<JsonException>();
+    }
+#endif
+
+#if TEST_SDK_NETSTANDARD21
+    [TestMethod]
+    public void IsoDateStringJsonConverter_RejectsInvalidDateFormat_OnRead()
+    {
+        var json = """{"birthDate":"15-01-1990"}""";
+
+        var act = () => JsonSerializer.Deserialize<NaturalPersonKycFields>(json, KycJsonOptions.Default);
+
+        act.Should().Throw<JsonException>();
+    }
+
+    [TestMethod]
+    public void IsoDateStringJsonConverter_RejectsInvalidRegistrationDate_OnRead()
+    {
+        var json = """{"registrationDate":"June 1, 2010"}""";
+
+        var act = () => JsonSerializer.Deserialize<OrganizationKycFields>(json, KycJsonOptions.Default);
+
+        act.Should().Throw<JsonException>();
+    }
+
+    [TestMethod]
+    public void IsoDateStringJsonConverter_RejectsInvalidDateFormat_OnWrite()
+    {
+        var fields = new NaturalPersonKycFields { BirthDate = "6/9/2026" };
+
+        var act = () => JsonSerializer.Serialize(fields, KycJsonOptions.Default);
+
+        act.Should().Throw<JsonException>();
+    }
+
+    [TestMethod]
+    public void IsoDateStringJsonConverter_AllowsNullDates()
+    {
+        var fields = new NaturalPersonKycFields { FirstName = "John" };
+
+        var json = JsonSerializer.Serialize(fields, KycJsonOptions.Default);
+        var roundTrip = JsonSerializer.Deserialize<NaturalPersonKycFields>(json, KycJsonOptions.Default);
+
+        roundTrip!.BirthDate.Should().BeNull();
     }
 #endif
 }