+++ StellarDotnetSdk/Sep/Sep0009/OrganizationKycFields.cs
@@ -1,9 +1,7 @@
 using System;
 using System.Collections.Generic;
-#if !NETSTANDARD2_1
 using System.Text.Json.Serialization;
 using StellarDotnetSdk.Converters;
-#endif
 
 namespace StellarDotnetSdk.Sep.Sep0009;
 
@@ -140,6 +138,7 @@ public sealed record OrganizationKycFields
     ///     netstandard2.1.
     /// </remarks>
 #if NETSTANDARD2_1
+    [JsonConverter(typeof(IsoDateStringJsonConverter))]
     public string? RegistrationDate { get; init; }
 #else
     [JsonConverter(typeof(NullableDateOnlyJsonConverter))]